+++ plugins/IdeaVimToggleIME/CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ## [Unreleased]
 
+## [1.0.2] - 2026-03-02
+
+### Changed
+
+- update IdeaVIM compatibility to `2.30.0`
+- make IME listener lifecycle unload-safe to keep mode/focus synchronization stable
+
+### Added
+
+- add `ImeStateControllerTest` coverage for mode transitions and focus lifecycle behavior
+
 ## [1.0.1] - 2025-08-25
 
 ### Fixed
@@ -16,6 +27,7 @@
 
 - Auto toggle IME for IdeaVim
 
-[Unreleased]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ghostflyby/IntelliJ-Plugins/commits/v1.0.0